auth-service-python/requirements.txt
flask
flask-jwt-extended
flask-cors
python-dotenv
flask-sqlalchemy
PyJWT
Werkzeug